% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretiseNB.R
\name{fit.mdlp}
\alias{fit.mdlp}
\title{Discretise with MDLP method}
\usage{
fit.mdlp(formula, data)
}
\arguments{
\item{data}{}
}
\value{
A Dataframe with values discretised
}
\description{
Discretise with mdlp method using the library FSelectorRcpp
}
