% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_naive_baise.R
\name{fit}
\alias{fit}
\title{Naive Bayes Classifier Constructor}
\usage{
fit(formula,data, m=1, discretise="rpart", selectvar=TRUE)
}
\arguments{
\item{formula}{Formula to specifies which variables in data use to fit the model.}

\item{data}{A dataframe with the data..}

\item{m}{Integer to use the Laplacian in the calcul of probabilities.}

\item{discretise}{String to choose the discretising method. Can be "rpart" (recommanded) or "MDLP".}

\item{selectvar}{Boolean to enable/disable variables selection (stepforward).}
}
\value{
A NBAYES object (S3)
}
\description{
Build a Naive Bayes Classifier model from a dataframe. It discretise variables, compute conditionnal probabilities and select variables if choosen.
}
\examples{
data(iris)
df = iris
f = Species ~ .
df$Species = as.factor(df$Species)
bayes = fit(f, df, discretise = "rpart")
}
