% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretiseNB.R
\name{fit.rpart}
\alias{fit.rpart}
\title{Discretise with rpart method}
\usage{
fit.rpart(Y, quanti)
}
\arguments{
\item{quanti}{}
}
\value{
A Dataframe with values discretised
}
\description{
Discretise with rpart method using the library rpart
}
