% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methodsNB.R
\name{plot.NBAYES}
\alias{plot.NBAYES}
\title{Summary from NBAYES object}
\usage{
\method{plot}{NBAYES}(objet)
}
\arguments{
\item{objet}{}
}
\value{

}
\description{
Plot the normalized statistics of the link between Y an each variable
}
