% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_naive_baise.R
\name{predict.NBAYES}
\alias{predict.NBAYES}
\title{Predict from NBAYES}
\usage{
predict(object, newdata, type="class")
}
\arguments{
\item{object}{Object of classe NBAYES fitted model}

\item{newdata}{Dataframe with data to classify. All the variables used by the model (see NBAYES$X) have to be present in the dataframe.}

\item{type}{String to choose the format of the output. Can be "class" to only have predicted classes, or "posterior" to have the posterior probability for each class.}
}
\value{
A list with the predictions.
}
\description{
Classify new data from a NBAYES model
}
\examples{
data(iris)
df = iris
f = Species ~ .
df$Species = as.factor(df$Species)
bayes = fit(f, df, discretise = "rpart")
pred = predict(bayes, df)
}
