% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methodsNB.R
\name{print.NBAYES}
\alias{print.NBAYES}
\title{Print the NBAYES object}
\usage{
\method{print}{NBAYES}(objet)
}
\arguments{
\item{objet}{The NBAYES object returned by the fit function}
}
\value{
Print some attributs of the NBAYES objects :

Discretisation method used for the fit (disc.param)

Varriable to classify (Y)

Variables used by the model - only variables kept after discretisation and selection (X)

Prior probabilities computed (apriori)

COnditional probabilities (p_conditional)
}
\description{
Print the NBAYES object
}
