% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methodsNB.R
\name{summary.NBAYES}
\alias{summary.NBAYES}
\title{Summary from NBAYES object}
\usage{
\method{summary}{NBAYES}(objet)
}
\arguments{
\item{objet}{}
}
\value{
Print the significance of the link between each variable X and Y (signif)
}
\description{
Summary from NBAYES object
}
