% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretiseNB.R
\name{transform.mdlp}
\alias{transform.mdlp}
\title{Discretise newdata by using the MDLP model}
\usage{
\method{transform}{mdlp}(intervals, values)
}
\arguments{
\item{intervals}{Dataframe of discretised variables by the fit.mdlp function (as a model).}

\item{values}{New Dataframe to discretise. All the columns have to be present in the intervals parameter.}
}
\value{
A Dataframe with values discretised
}
\description{
Discretise newdata by using the MDLP discretisation retuned by fit.mdlp
}
