% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretiseNB.R
\name{transform.rpart}
\alias{transform.rpart}
\title{Discretise newdata by using the rpart model}
\usage{
\method{transform}{rpart}(objetRpart, newdata)
}
\arguments{
\item{objetRpart}{The Rpart object returned by fit.rpart (model)}

\item{newdata}{New data to discretise}
}
\value{
Dataframe of discretised data
}
\description{
Discretise newdata by using the rpart model
}
