% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionvarNB.R
\name{selection_}
\alias{selection_}
\title{Step-forward variables selection}
\usage{
selection_(y, X)
}
\arguments{
\item{y}{classe to predict}

\item{X}{set of variables to choose}
}
\value{
A list with the name of selected columns
}
\description{
Methode used by fit.NBAYES to select variables.

It is a step-forward algorithme.
}
