% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionvarNB.R
\name{signification_}
\alias{signification_}
\title{Compute the significance of the link between each variable X and Y}
\usage{
signification_(y, X)
}
\arguments{
\item{X}{}
}
\value{

}
\description{
Compute the significance of the link between each variable X and Y
}
