% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretiseNB.R
\name{splitdf}
\alias{splitdf}
\title{Split a dataframe by variables classes}
\usage{
splitdf(formula, data)
}
\arguments{
\item{data}{}
}
\value{
A list with 3 dataframes : quali, quanti and Y
}
\description{
Split a DataFrame in 3 part : categorial (quali), numeric (quanti) and the variable to predict(Y) if present
}
